requirements.txt
strawberry-graphql==0.205.0
plaid-python==16.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2 